introduce event-listeners
another approach to visualizing intermediary state.
Instead of having RendererAtoms that you add to the graph,
I introduce event-listeners, that can bring state out of the graph.

With `ctx.event.addListener(nodeId, (value: any) => ...)` you can add
a listener. It returns an unsubscribe-function.

Currently the emiting of events is opt-in:
When marking a node with `emits: true`, we wrap its function-call with
a `ctx.event.emits(nodeId, ...)`.
This function then triggers the subscribed listeners.

reflection:
To allow for any function-call to listen to, `ctx.lib` could be a proxy.
There is a performance-cost with proxies, which is why I am
first looking for alternatives.

I am unsure if the events eventually need to be scoped to ctx,
I am unsure if compiling the emits is the best way either
both could be revisited in a later commit.